arm64: dts: sdmmagpie: Remove display ramdump memory region
This region is the same as the early splash framebuffer initialized by
XBL. msm_drm uses it for ramdump when available, but that's a debugging
feature we don't need. It also wastes memory after early boot because
msm_drm will keep it reserved rather than freeing the framebuffer memory
once it has taken control of the display hardware.

Remove the display ramdump region to save 36 MiB of memory and silence
the following error on boot:

[    0.000000] OF: reserved mem: OVERLAP DETECTED!\x0adisp_rdump_region (0x000000009c000000--0x000000009e400000) overlaps with cont_splash_region (0x000000009c000000--0x000000009e400000)

Change-Id: I0b7a1c87d5a449e9fd19b770497e5c3f7a62c6d6
Signed-off-by: Danny Lin <danny@kdrag0n.dev>